Locale objects use atomic reference counting, which may be very expensive in parallel applications. The classic locale is used by default by all streams and can be very contended. But it's never destroyed, so the reference counting is also completely pointless on the classic locale. Currently ~70% of time in the parallel stringstream benchmarks is spent in locale ctor/dtor. And the execution radically slows down with more threads.

Avoid reference counting on the classic locale and inline common ctors/dtor. With this change locale ctor/dtor time become negligible and the benchmark starts to scale with threads.

```
                                       │ bench/stream.baseline.res │     bench/stream.optimized.res      │
                                       │          sec/op           │   sec/op     vs base                │
Istream_numbers/0/real_time/threads:1                  4.686µ ± 0%   4.213µ ± 0%  -10.09% (p=0.000 n=24)
Istream_numbers/0/real_time/threads:72               339.712µ ± 4%   8.939µ ± 2%  -97.37% (p=0.000 n=24)
Istream_numbers/1/real_time/threads:1                  4.900µ ± 0%   4.773µ ± 0%   -2.60% (p=0.000 n=24)
Istream_numbers/1/real_time/threads:72                49.308µ ± 3%   9.694µ ± 1%  -80.34% (p=0.000 n=24)
Istream_numbers/2/real_time/threads:1                  4.902µ ± 0%   4.769µ ± 0%   -2.72% (p=0.000 n=24)
Istream_numbers/2/real_time/threads:72                 416.6µ ± 9%   410.5µ ± 6%        ~ (p=0.111 n=24)
Istream_numbers/3/real_time/threads:1                  4.716µ ± 0%   4.737µ ± 0%   +0.45% (p=0.000 n=24)
Istream_numbers/3/real_time/threads:72                 419.7µ ± 6%   458.3µ ± 4%   +9.19% (p=0.003 n=24)
Ostream_number/0/real_time/threads:1                   184.0n ± 1%   135.0n ± 1%  -26.63% (p=0.000 n=24)
Ostream_number/0/real_time/threads:72                17279.0n ± 4%   308.0n ± 1%  -98.22% (p=0.000 n=24)
Ostream_number/1/real_time/threads:1                   253.5n ± 1%   199.0n ± 1%  -21.50% (p=0.000 n=24)
Ostream_number/1/real_time/threads:72                13957.5n ± 2%   382.0n ± 0%  -97.26% (p=0.000 n=24)
Ostream_number/2/real_time/threads:1                   253.0n ± 1%   202.0n ± 1%  -20.16% (p=0.000 n=24)
Ostream_number/2/real_time/threads:72                  28.04µ ± 4%   18.80µ ± 7%  -32.96% (p=0.000 n=24)
Ostream_number/3/real_time/threads:1                   187.0n ± 1%   190.5n ± 0%   +1.87% (p=0.000 n=24)
Ostream_number/3/real_time/threads:72                  20.25µ ± 7%   20.37µ ± 5%        ~ (p=0.736 n=24)

```