@philnik777 2023-10-30
The numbers look really good, but as-is this is almost certainly ABI-breaking. You can't just move functions to and from the dylib. From taking a quick look at the changes, I think this can be done without breaking the ABI. I'd also like to see benchmarks for other locales, since they will probably be affected negatively by this change.