@firewave 2023-10-30
See #56202 for a related issue which also contains some benchmark code which might be covering what @philnik777 requested .