@dvyukov 2023-10-31
> I think this can be done without breaking the ABI

How? Do you have any specific way in mind? Are there examples of such changes?