@firewave 2023-10-31
> > See #56202 for a related issue which also contains some benchmark code which might be covering what @philnik777 requested .
> 
> I don't see any benchmarks there, nor in referenced changes. Which ones do you mean?

https://github.com/llvm/llvm-project/issues/56202#issuecomment-1166229891