@dvyukov 2023-10-31
I've extended benchmarks to test non-classic locale, and update benchmark results in the PR description.