@dvyukov 2023-10-31
I've added _LIBCPP_HIDE_FROM_ABI to inline functions as the bot suggested.

Now some tests are failing with dynamic libc++.so, e.g.:
```
# | t.tmp.exe: libcxx/test/std/localization/locales/locale/locale.cons/default.pass.cpp:81: int main(int, char **): Assertion `loc2 == std::locale(LOCALE_en_US_UTF_8)' failed.
```

Is it possible to implement new inline functions in libc++?